fix the loading API and test it with the minko player add the corresponding Lua bindings for the new classes
Former-commit-id: c9c9a2beb679f4e2df288726dd6a98e7950a4a96

+++ plugin/http-worker/src/minko/async/HTTPWorker.cpp
@@ -111,7 +111,7 @@ namespace minko
 
 			double ratio = current / total;
 
-			worker->progress(ratio);
+			worker->progress((float)ratio);
 
 			return 0;
 		}
